project-bolt-sb1-bkdtquqs/project/src/backend/requirements.txt
# backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
Pillow==10.0.1
python-dotenv==1.0.0
gunicorn==21.2.0

# Optional: For production deployment
# redis==5.0.1
# celery==5.3.4

# Optional: For database integration
# Flask-SQLAlchemy==3.0.5
# Flask-Migrate==4.0.5
# psycopg2-binary==2.9.7  # For PostgreSQL
# pymongo==4.5.0          # For MongoDB

# Optional: For AI integration
# openai==0.28.1
# replicate==0.12.0
# requests==2.31.0